Update project objectives and simplify code in simplNotas.py
The project's README.md file has been updated to reflect the completion of a few tasks: calculating weighted average of students and reading the attendance list. On the other hand, unnecessary variables were removed and function names were simplified in the simplNotas.py file to improve readability. The code at script-end was also updated to print the sorted list of students based on their scores.

Signed-off-by: AliCCFire <alicecfire@gmail.com>