Refactor simplNotas.py for readability and improved code execution
Renamed the 'simplNotas' class to 'SimplifiedGrades' and 'get_media' method to 'get_avg_grade' for better clarity and readability. Regardless of whether a string path or dataframe was used to initialise grades or absences data, a more flexible and cleaner handling of data types was implemented. The method 'rendimento' was refined to 'get_performance' and now returns a 'category' data type for efficient memory usage.

Signed-off-by: AliCCFire <alicecfire@gmail.com>